+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- mobile: add ability to edit tags
 - replace dive computer management with a new implementation, this fixes rare cases where
   Subsurface unnecessarily download already downloaded dives, but also means
   that some older style dive computer can no longer have nicknames